+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to remove an Identifier of an (inherited) `Itendifiable`
 - Builder for `Identifier`
 
+### Removed
+
+- **Breaking** Removed obsolete fields `language`, `publicationDat`,
+  `publicationPlace`, `publisher` and `version` in `Item` since they
+  belong to the Manifestation/Expression.
+
 ## [10.0.2](https://github.com/dbmdz/digitalcollections-model/releases/tag/10.0.2) - 2022-06-21
 
 ### Removed